gspc: new: implemented configurable heartbeats
A connection  contains two heartbeat  timers, a send  and a receive  timer. That
means, it  will send heartbeats with  a given interval  (or none at all)  and it
will expect heartbeats from the other side in a given interval (or none at all).

When the  connection did not receive  a heartbeat after the  given duration, the
connection will be considered dead.