rpc: add exception support.
throwing exceptions requires the exception type to be
serializable. std::* exception classes have builtin serialization. for
custom types, serialization & deserialization functions need to be
supplied (see remote_endpoint and service_dispatcher). if no function is
supplied, it is tried to fall back to the nearest std::* exception class
if inherited from one. if not, std::terminate() is called. if there is
only a serialization but no de-serialization function, it will be tried
be downcast to a std::* exception again, if not possible
std::terminate() is called.

serialization needs to convert a std::exception_ptr to an std::string,
if possible, or return boost::none if exception is unknown.

deserialization needs to be provided in two forms: returning a
std::exception_ptr for the previously returned string, and calling
std::throw_with_nested() with an exception constructed from the
previously returned string.

functions are identified by a name. on deserialization side, the
function of same name is called. if multiple serialization functions
would not return none, it is undefined which one is called, but
guaranteed that the matching deserialization is called.

std::nested_exception is supported.

note that std::regex_error, std::ios_base::failure and
std::io_stream_category are crippled due to bugs in libstdcxx of gcc-4.8

struct example : std::runtime_error
{
  using std::runtime_error::runtime_error;

  static std::exception_ptr to_exception_ptr (std::string what)
  {
    return std::make_exception_ptr (example (what));
  }
  static void throw_with_nested (std::string what)
  {
    std::throw_with_nested (example (what));
  }
  static boost::optional<std::string> from_exception_ptr (std::exception_ptr ex)
  {
    try
    {
      std::rethrow_exception (ex);
    }
    catch (example const& e)
    {
      return e.what();
    }
    catch (...)
    {
      return boost::none;
    }
  }
};

service_dispatcher ({{"example", &example::from_exception_ptr}})
remote_endpoint (…, {{"example", &example::to_exception_ptr, &example::throw_with_nested}})