vmem: hack/fix: ensure that write ids only span a single byte
workaround for rui.machado/gpi-2#10: notification values are not written
atomically in gpi-2+eth but on a byte-by-byte level, thus we limit us to
using just a single byte at any given time. fixing the issue means that
we can use the whole range of notification_t again. note that the
threaded queue could possibly be optimized to a pair of "next" and "last
freed", but the check for wrapping still needs to exist as long as
write_id_max < (ping_ids_per_node * num_ranks), which currently is the
case quite quick as 256*4 < 8 * r only for r < 128.