vmem: revert 36da56659: memory_size at this point is per node
from the comments on gitlab:

it appears that I confused myself with 2ddf5e38 where the global
allocation with --vmem-size-per-node removed and replaced with
drts.alloc() which takes the global size. This global size is there
used to just-in-time create the segment, with
pc::client::api::create_segment -> memory::manager_t::add_memory ->
either sfs_area::create or gpi_area::create.

In sfs_area_t::sfs_area_t there is a little comment "// total", which
from a quick look seems to be correct (the file is truncated to that
size, there is only one file).

In gpi_area_t, this size is passed up into gpi::api::gaspi_t where it
is passed into gaspi_segment_create and thus interpreted as a local
size (it is also compared to the local memory available, so that
matches).

Both times, the size given is the one used for area_t::m_descriptor
where size suddenly becomes local_size in the ctor argument list and
the notion of a global size vanishes.

Thus, it appears that since 2ddf5e38, the gaspi segments will be
num_node times as large as the total size requested, while filesystem
segments are fine. The communication buffer space was correct as it
was using per_node_size = total_size + com_size * num_com. The change
that should have been in this commit is thus, that (at a different
position though), the size passed in gpi_area_t::create is divided by
the node count, thus making it per_node_size = (total_size /
num_nodes) + com_size * num_com.