ci: re-build libssh2 with fpic to avoid having to fix cmake
rpath handling of imported libraries seems broken, but the only
alternative would be to fall back to having non-absolute paths in
linking (which is an issue for libraries in different versions than the
"system" ones which cmake automagically detects (e.g. libdl in
/usr/lib64/ and /lib64). The only other way to get both, rpath and
absolute paths, seems to be `link_directories` which is explicitly
marked as to be avoided and can not be specified on a target basis but
only for a whole directory, which one obviously does not
want. Following, see the linker command line produced for different
ways of linking, assuming

 add_library (foo SHARED IMPORTED)
 set_property (TARGET foo PROPERTY IMPORTED_LOCATION ${foo_path})

> target_link_library (bar foo)

 /home/projects/sdpa/external/libssh2/1.6.0/lib/libssh2.so
 /usr/lib64/libdl.so
 -Wl,-rpath,/home/projects/sdpa/external/boost/1.59.0/lib

> target_link_library (bar ${foo_path})

 /home/projects/sdpa/external/libssh2/1.6.0/lib/libssh2.so
 -ldl
 -Wl,-rpath,/home/projects/sdpa/external/boost/1.59.0/lib:/home/projects/sdpa/external/libssh2/1.6.0/lib:

 $ ldd … | grep libdl
 libdl.so.2 => /lib64/libdl.so.2 (0x000000356ba00000)

> target_link_library (bar foo ${foo_path})

 /home/projects/sdpa/external/libssh2/1.6.0/lib/libssh2.so
 /home/projects/sdpa/external/libssh2/1.6.0/lib/libssh2.so
 /usr/lib64/libdl.so
 -ldl
 -Wl,-rpath,/home/projects/sdpa/external/boost/1.59.0/lib:/home/projects/sdpa/external/libssh2/1.6.0/lib:

 $ ldd … | grep libdl
 libdl.so.2 => /lib64/libdl.so.2 (0x000000356ba00000)

> link_directories (/home/projects/sdpa/external/libssh2/1.6.0/lib/)
> target_link_library (bar foo)

 -L/home/projects/sdpa/external/libssh2/1.6.0/lib
 /home/projects/sdpa/external/libssh2/1.6.0/lib/libssh2.so
 /usr/lib64/libdl.so
 -Wl,-rpath,/home/projects/sdpa/external/libssh2/1.6.0/lib:/home/projects/sdpa/external/boost/1.59.0/lib:

Where none of them fits our goal of

 /home/projects/sdpa/external/libssh2/1.6.0/lib/libssh2.so
 /usr/lib64/libdl.so
 -Wl,-rpath,/home/projects/sdpa/external/libssh2/1.6.0/lib:/home/projects/sdpa/external/boost/1.59.0/lib:

Thus, in order of spending even more time reading cmake code and
googling, I decided to postpone it (to probably never) and just manually
install the dependencies in a form where no rpath handling should be
required due to being all static. Wohoo!