drts: add test for steal_work not holding a reference to a popped queue element
A reference to the top of a queue was held, which was then popped but
conditionally used after popping again. The condition was that a worker
may be stolen from again. This means that it has a running and a pending
job, or multiple pending jobs, after stealing once.

This can be trivially triggered by having three jobs which are all
assigned to the same worker, and two workers without jobs. Both workers
will then try to steal from the first one, which results in undefined
behaviour as the reference will point to an element in an empty queue
when stealing the second time.

If now a fourth worker exists which can also be stolen from, instead of
pointing into an empty queue, it will be pointing to a different element
of the queue, which leads to emplacing a different element of the queue
a second time, rather than emplacing the intended worker. As this is
easier to detect than an invalid-read, this test case only does this
test.