rif: fix: on bootstrapping, remove failures from registered entry_points
A bootstrapping may have properly run, but after bootstrapping, the
strategy may have decided the node failed, e.g. due to output on stderr
in the ssh strategy. In that case, we should for safety assume the node
as broken, rather than as bootstrapped. This fixes an endless wait.

Note: this might leak rifds.