no longer define BOOST_VARIANT_DO_NOT_USE_VARIADIC_TEMPLATES
This has to be done since both INTERTWinE and GPI-Space use
boost.variant, but INTERTWinE is using the variadic version. The
alternative would have been to use a selfwritten variant, or to make
them compatible with section wise definition of the macro, and other
pain, which is both not favorable.

This means that our clients must all be using C++11 now. To do so in our
tests, remove the gen::cxx11() flag and add it to every command line.