fhgcom: test: don't use port 1235 but 12351
We hardcode the port to be able to check that a port is used if we
specify one. The port years ago chosen by dice roll is used on seislab
though to monitor if a node is up:

    $ nc node45 1235
    UP

Thus, roll the dice again and choose 12351, which I checked to not be
in use on seislab, build-centos*, beehive, lts.

It will probably clash again in another five years, but at least until
then, the test will run on all hosts we know.