testing: add hopefully_free_port
When needing fixed ports for tests, actually hardcoding them would be
error prone. This utility constructs a TCP RPC provider with automatic
port assignment, which then is destructed upon `release()` right
before creating the real resource with the port, in the hope that no
other thread or process gets the port assigned in the meantime.

Note: This may still end in failed tests! The ports are *hopefully*
free, not guaranteed.

Note: This is probably better put into util-generic, but it uses RPC,
so that would be a circular dependency. Thus, this helper exists in
copies in multiple projects. One other instance is in INTERTWinE-vmem.