logging: rename emit to emit_message
Qt has `#define emit` globally (unless explicitly requesting it not
to, which we don't), which makes having any symbol with the name
`emit` impossible after including Qt. Note that this only needs to be
the case in headers since logging itself does not include Qt, so there
are some unchanged occurrences.