monitor: add ability to receive gantt via new style tcp emitters
Currently, this is a pure addition, with the whole UDP based
infrastructure still existing. An application can now publish the log
emitters to the user and have the user start a monitor listening to
them.

An application is likely going to want to aggregate the log emitters
before exposing to the user to avoid having the user specify hundreds
of endpoints. That would also allow that aggregation server to use a
fixed port again so that a monitor can be reused across jobs, assuming
the monitor will reconnect.