we: plugins: Log
provides a minimal plugin to log the evaluation context into a file and
demonstrate how to use it in a Petri net.

Note: The evaluation context stores values in unordered structures. So
the plugin order and the test program order might differ. The tests take
that into account when comparing results.