logging: use only new style logging for gantt
This implies that `--gui-port` and `--gui-host` are removed. Instead,
when (currently always) a logging rif entrypoint is given,
`scoped_runtime_system` will output a logging endpoint on startup:

```
I: starting top level gspc logging demultiplexer on lts003hpc.itwm.fhg.de
   => accepting registration on 'TCP: <<lts003hpc.itwm.fhg.de:41152>>, SOCKET: <<lts003hpc.itwm.fhg.de:\0ae9c7>>'
```

This endpoint can then be used to pass to gspc-monitor:

```
/dev/shm/loerwald/gpispace/install/bin/gspc-monitor \
  --log-port 11042 \
  --emitters 'SOCKET: <<lts003hpc.itwm.fhg.de:\0ae9c7>>'
```

Note: `--emitters` could be either

* `TCP: <<lts003hpc.itwm.fhg.de:41152>>, SOCKET: <<lts003hpc.itwm.fhg.de:\0ae9c7>>`
* `SOCKET: <<lts003hpc.itwm.fhg.de:\0ae9c7>>` (only if on lts003hpc)
* `TCP: <<lts003hpc.itwm.fhg.de:41152>>`, or
* `SOCKET: <<lts003hpc.itwm.fhg.de:\0ae9c7>>, TCP: <<lts003hpc.itwm.fhg.de:41152>>`

which are all equivalent. Not being able to specify the port to use on
startup anymore is to be re-introduced with being able to specify
demultiplexer rif entry point.