drts: drop old style logger from genericdaemon and thus tests
Note: This means that tests are now nearly silent, which is not a good
thing. We should^tm resurrect logging to console at some point, but
with a fixated *stream* emitter everywhere, that's a pain. I think it
is a good idea to make an abstract emitter base class and just use a
stdout emitter in tests etc.