drts: test: bump randomly chosen timeout for wait_for_capabilities calls
With SSL versions of the tests, the timeout is very close to the time
needed, making the tests fail when run in parallel. As this is not a
performance modeled timeout but just one to avoid the test hanging
forever, bump that timeout to allow tests passing on buildbots with
parallel tests.