we: net: test: partial cross product for conditions
The test creates a network

       (x::bool) -> [ t        ]
    (y::control) -> [ if: ${x} ]

and puts `N` tokens on place `y` before putting a single token with
value `false` on place `x`. This last put triggers the evaluation of the
condition which not satisfied and should be done in constant time rather
than `O(N)`.

Tests #795

Note: Fails badly because the ratio between fastest and slowest
execution is O(500000) rather the required 100.