test: add check that all installed headers are selfcontained
Only allowed includes are from system, our bundled boost, and our
installed headers.

This is not perfect: Note that the system part of inclusion means that
the test runner needs to be as clean as possible, as e.g. on lts we
don't even need to include our boost as it is globally installed. The
alternative would be to pass `-nostdinc', but that disables absolutely
everything, including STL headers, which we don't want to fake or
manually add again.

Note: test currently fails due to

- logging/*endpoint* being missing and requiring util-generic
- rif/started_process_promise.ipp being missing
- stencil_cache requiring util-generic