stencil_cache: fix destruction order
The allocation thread executes

    _scache.allocate (_queue_allocate.get());

so it might be inside the `allocate` when the desctructor (e.g. due to
cancel/abort/failure) gets called. To interrupt the queue does not
prevent the allocate to wait for additional free slots...

Both operations must be interrupted before joining the thread. To
interrupt the queue first is important for the case where the thread is
inside the `allocate`.

- fixes ae1e667c43868456db7c38132ddad7fe71a978ed

- Closes #799