drts, logging: restore ability to have automatically log to GUI
Since new logging requires active subscription of the GUI with the log
emitter, specifying `--log-host` and `--log-port` as before was no
longer possible.

Now, `gspc-monitor` learned `--port` again, which can then be given to
the runtime system to have the runtime system tell the monitor on that
port to subscribe to the top level emitter.

Note that this is an optional argument and the top level emitter
endpoint is still printed on startup regardless of arguments, to allow
for later re-attachment without drts-restart.