stencil_cache: test: random sleep
The (correct) implementation reads

```
    stencil_cache::~stencil_cache()
    {
      _queue_allocate.interrupt();
      _scache.interrupt();
      _allocate.join();
    }

    void stencil_cache::allocate()
    try
    {
      while (true)
      {
        _scache.alloc (_queue_allocate.get());
      }
    }
    catch (QAllocate::interrupted)
    {
      return;
    }
    catch (SCache::interrupted)
    {
      return;
    }
```

and all implementations would be wrong where any of the two `interrupt`
comes after the `join`.

There is no way to reliable test this condition though. To sleep for a
random duration mimics the behavior of an application that executes
_some_ work between a call to `alloc` and the (out of sync) destruction
of the object, e.g. due to an error somewhere else.