vmem: allow specifying --virtual-memory-netdev-id, passed to GPI
Due to rui.machado/GPI-2#31 customers that have two devices but the
defaulted one is requiring RoCE are unable to run GPI-Space with
virtual memory.

This diff allows them to overwrite auto-detection, which does not fix
the broken RoCE support, but allows them to use the Infiniband device.

Note: No system test because we do not actually run CI on machines
with IB, and the option does nothing with Ethernet.