we: test: fix operator equal
- fixes 88bf8ce028d92bdf1e9fe646977bd24bf7f11a4f and
2b688266c18df48b774800653ea038cafa599f79:

In 2b688266c18df48b774800653ea038cafa599f79 the `eureaka_id` was added
but not taken into account in the `operator==`. Which was not a
problem at that time as the `operator== (acticity_t)` did use the
`transition_id()` as a tie breaker even when the `eureka_id` would
have been the correct tie breaker.

Later, in 88bf8ce028d92bdf1e9fe646977bd24bf7f11a4f, the
`transition_id` oberserver got removed by a line of argumentation that
states it would bot be relevant. However, in the `eureka`-test the
used transitions are equal except for the `eureka_id` and the
`transition_id`.