fhgcom: test: add test for hangsegfaultcrashcorruptdata in ~peer_t
A race in destruction of peer_t, while a message was just sent,
resulted in

- corrupted messages being received (assertion fails)
- segfaults in various places (e.g. SSL context destruction)
  (signal raised)
- 'header/data length mismatch' when the destructing peer was still
  trying to send (terminate due to being thrown in some thread)

All of the above are not easily triggered by anything in
particular, so this test tries to reproduce them by doing the
construct-connect-send-destruct sequence multiple times. The
repetition counter was adjusted to take less than 30 seconds on LTS
while still reproducing every few runs.