fhgcom: fix: only close socket in dtor, not stop
Do not close before all async operations are done. This includes
things posted onto the strand without mentioning the socket, which
are *not* canceled when _socket.cancel() is called and would likely
segfault. (Segfault seen is `descriptor_data->shutdown_`, which is set
to 0 on close.)

By only closing in the dtor, those operations will see a shutdown, but
not closed socket, and after failing to do whatever they want to do
release their reference to connection_t one after the other.

All current calls to strand have been checked and have a shared_ptr to
`this` and/or reference the socket and will thus be canceled.