drts: test: remove randomness for sake of clarity of test
The test checks that a blocking workflow response `check_done` will
block fine while a workflow response `get_and_update_state` is
executed multiple times. In addition to that, there is a randomly
sleeping module call that once done lets `check_done` return.

Before this change, that randomly sleeping module call has 20-50
invocations that slept 10-50ms each. Now there is one invocation that
sleeps 200-2500ms. This is technically no difference (other than that
the agent has to submit more tasks, which shouldn't be relevant for
the goal of the test?!), it is still just as random how long/often
`get_and_update_state` is called. There are still just as many
workflow responses in flight. It mostly removes overhead in
understanding what is even happening in the test.