drts: test: let use_fixed_ports retry in case of hitting already-in-use race
Even if we directly allocate a port, release it and let the DRTS use
it, there still is a race that the port could be already in use by a
different process again.

This is quite rare but for the sake of tests not spuriously failing in
this case, give this part of the test more than one try as it is not
what is wanted to test: We are interested if after startup *we* use
those ports, not if we fail starting up because someone else is using
them (which implicitly also confirms test success but just for one
component).