iml: test: do silent fallback to gspc nodefile until it is really separate
The only consequence from not doing this that every developer needs to
adapt right now, which includes users of a public release where we
only document the one environment variable...

A proper solution would wrap test execution in some cmake/shell script
that automatically sets the environment variable, rather than hacking
it inside the tests. In either case, a GPI-Space user/developer should
never know what an IML is and bother about that to test GPI-Space.