iml: implement decision to globally alloc/free only once
Done if the allocation is global (always true for BeeGFS, never for
shm, true for non-communication buffers for Gaspi case), and for alloc
also only if creator is valid (i.e. not a follow up of a remote
alloc).

Note: Not having the loop detection on free is weird, but not a
change. This works as `remote_free()` does not call `free()` again,
but re-implements it.