sdpa: test: fix: don't send a default constructed activity on finish
When injecting the (empty) result activity, `_transition_id` (an
optional) is dereferenced based on the assumption that every job sent
out has one, so every job received back should have one. This was an
invalid read in Release builds and an assertion in Debug builds.

By returning the input activity, it is ensured it is a valid activity
as expected by the layer. This transfers useless data, but that's also
the case in the real worker, so the test doesn't have to optimize that.