release 22.03
## Multi Scheduler Support

GPI-Space is adding support for using multiple schedulers.
The scheduler selection is automatically performed by the agent upon workflow submission.
At that time, the workflow's properties are analyzed and an appropriate scheduler is chosen accordingly.

## Customizable Number of Task Retries

A user is now allowed to specify an upper limit for the number of task retries in case of worker failures.
This can be done by setting the `maximum_number_of_retries` scheduling property in the workflow, as in the example below:

```xml
<transition name="transition">
  <defun name="transition">
    <properties name="fhg">
      <properties name="drts">
        <properties name="schedule">
          <property key="maximum_number_of_retries">"5UL"</property>
        </properties>
      </properties>
    </properties>

    ...

    <module name="module" function="f (input, ...)">
      <code><![CDATA[
      ...
      ]]></code>
    </module>
  </defun>

  ...

</transition>
```

If this property is not explicitly set, it falls back to its default behavior.
Which means a task is rescheduled and retried infinitely in case of repeated worker failures.

## Miscellaneous

- CentOS 8 has reached its End Of Life on December 31th 2021.
  For that reason GPI-Space is dropping support for CentOS 8 and replacing it with Oracle Linux 8.
- The co-allocation scheduler's performance has been improved.
  Workflows containing tasks with a multiple workers requirement benefit from a significant speedup.