release 22.09
## Optional Components: IML

The Independent Memory Layer (`IML`) is an abstraction layer for easy distributed memory management, allowing applications to cache data, or to prepare and extract data independent of GPI-Space across runs.
GPI-Space now supports disabling the `IML` at configuration time to reduce the dependencies needed as
well as build time.
This change makes `GPI-2` an optional dependency of GPI-Space.
The following options can be given to CMake or Spack to enable or disable the IML:

| Spack | CMake |
| - | - |
| `[+\|~]iml` | `-D GSPC_WITH_IML=[ON\|OFF]` |

By default the IML is enabled for backward compatibility.